refactor(direct-control): extract command result parser
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves the shared command-result JSON payload parser into src/session/command-result.ts while preserving host/port/state context merging, invalid JSON message/classification/details, and existing facade dependency injection for wrapper modules.

Focused session proof now covers successful command payload context merging and malformed JSON details. This keeps the current hand-wired modularization shape; future Effect/Bun resource/layer/schedule composition and Effect/oRPC procedure-core schemas remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, Effect/oRPC, or procedure-core proof claimed.
