+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -676,6 +676,13 @@ runtime/direct-control claims.
         telemetry/correlation contracts, procedure schemas, semantic CLI, AI
         ingestion, hotseat runtime proof, Effect/oRPC procedure-core work, and
         Task 2.9.4 matrix-row acceptance pending.
+  - [x] 4.10.18 Extract command-result JSON payload parser into a session owner
+        while preserving host/port/state merge behavior, invalid JSON
+        message/classification/details, and existing facade dependency
+        injection, and leaving validation helpers, command serialization,
+        telemetry/correlation contracts, procedure schemas, semantic CLI, AI
+        ingestion, hotseat runtime proof, Effect/oRPC procedure-core work, and
+        Task 2.9.4 matrix-row acceptance pending.
 - [x] 4.11 Extract map/visibility/GameInfo read atoms.
   - [x] 4.11.1 Extract map summary, plot snapshot, and map grid read
         wrapper/source owner while keeping the public facade export surface in
